Fix key for error store
The issue here was that when a validation error happens using this
validator the error keys that are getting stored was the primary attribute
while the nested attributes are stored on the message.

But because of the implementation of the nested validations error keys for
nested validations are stored in a nesting hash that is converted to a string.